订单页

+++ app/src/main/java/com/qi/market/module/shoppingcart/db/dao/ShoppingCartDao.kt
@@ -39,7 +39,7 @@ class ShoppingCartDao(context: Context) {
         }
     }
 
-    fun delete(id: Int) {
+    fun delete(id: Long) {
         var db = mHelper.writableDatabase
         db.beginTransaction()
         try {
